+++ src/components/CardProductos/CardPoductos.jsx
@@ -70,16 +70,15 @@ const CardProductos = ({
 
   const getRandomColor = (name) => {
     const colors = [
-        "#F44336", "#2196F3", "#4CAF50", "#FF9800", "#9C27B0", "#3F51B5",
-        "#E91E63", "#00BCD4", "#8BC34A", "#FFC107", "#673AB7", "#607D8B",
-        "#FFEB3B", "#795548", "#CDDC39", "#FF5722", "#9E9E9E", "#000000",
-        "#1B5E20", "#B71C1C", "#4A148C", "#1A237E", "#004D40", "#880E4F",
-        "#3E2723", "#212121", "#FFD600", "#00C853", "#C51162", "#AA00FF",
-        "#6200EA", "#304FFE", "#2962FF", "#00BFA5", "#64DD17", "#AEEA00",
-        "#FF6D00", "#DD2C00", "#37474F", "#455A64", "#546E7A", "#1DE9B6",
-        "#00E5FF", "#18FFFF", "#84FFFF", "#A7FFEB", "#B9F6CA", "#CCFF90",
-        "#F4FF81", "#FFFF8D", "#FFE57F", "#FFD740", "#FFAB40", "#FF6E40"
-      ];
+      "#F44336", "#2196F3", "#4CAF50", "#FF9800", "#9C27B0", "#3F51B5",
+      "#E91E63", "#00BCD4", "#8BC34A", "#FFC107", "#673AB7", "#607D8B",
+      "#C2185B", "#512DA8", "#388E3C", "#F57C00", "#7B1FA2", "#1976D2",
+      "#689F38", "#FBC02D", "#0288D1", "#D32F2F", "#388E3C", "#FFA000",
+      "#5D4037", "#303F9F", "#0288D1", "#C62828", "#1565C0", "#AD1457",
+      "#FF7043", "#558B2F", "#FFCA28", "#E64A19", "#5E35B1", "#0288D1",
+      "#283593", "#00796B", "#F06292", "#FF9800", "#6D4C41", "#1B5E20",
+      "#827717", "#8E24AA",
+    ];
     const index = name.length % colors.length;
     return colors[index];
   };
