+++ src/pages/ProductosPage/ManageProducts/ManageProductsUtils.js
@@ -1,5 +1,4 @@
-import { useEffect, useState } from "react";
-
+import { useEffect, useState, useMemo} from "react";
 
 /* Consulta interna par la pagina de roles Busqueda de usuarios*/
 export const useSerchPrductos = (productos) => {
@@ -30,4 +29,23 @@ export const useSerchPrductos = (productos) => {
     showNoResults,
     handleSearch,
   };
+};
+
+export const useCategoriasYFiltrado = (productos, filteredProductos) => {
+  const [selectedCategory, setSelectedCategory] = useState("Todas las Categorias");
+
+  // Generar lista de categorías (incluyendo "Todas las Categorias")
+  const categorias = useMemo(() => {
+    const categoriasSet = new Set(productos.map((p) => p.nombreCategoria));
+    return ["Todas las Categorias", ...categoriasSet];
+  }, [productos]);
+
+  // Filtrar productos por categoría seleccionada
+  const filteredByCategory = useMemo(() => {
+    return selectedCategory === "Todas las Categorias"
+      ? filteredProductos
+      : filteredProductos.filter((p) => p.nombreCategoria === selectedCategory);
+  }, [filteredProductos, selectedCategory]);
+
+  return { categorias, filteredByCategory, selectedCategory, setSelectedCategory };
 };
